# docs/openapi.yaml
openapi: 3.1.0
info:
  title: llmproxy
  version: 1.0.0
  description: |
    Self-hosted, OpenAI-compatible LLM proxy. A single static Go binary that
    routes OpenAI-dialect requests to registered upstream providers, with
    per-user API keys, a curated model catalog and per-unit usage accounting.

    Surface areas:

    * `/v1` - OpenAI-compatible ingress. Request bodies are passed through to
      the upstream byte-for-byte at the field level, so any field OpenAI's
      chat/completions/embeddings endpoints accept is accepted here. The proxy
      itself only reads or rewrites four top-level fields: `model` (replaced
      with the upstream model name), `stream`, `stream_options`
      (`include_usage` is forced to `true` on streamed requests) and, for
      embeddings, `input` (batch-size cap only). Everything else, including
      nested message content, vision parts and tool calls, is forwarded as raw
      bytes and never inspected. Upstream responses, including upstream error
      bodies and status codes, pass through unchanged. The same routes are
      also served at the root without the `/v1` prefix for clients configured
      that way.
    * `/model/*` - model-management compatibility endpoints (`/model/new`,
      `/model/info`, `/model/delete`) for existing proxy management tooling.
    * `/my` - self-service key, relay-token and usage management for the
      calling principal.
    * `/transparent/anthropic/{token}` - transparent Anthropic relay. Any
...

...
servers:
...
security:
  - bearerApiKey: []
  - apiKeyHeader: []
  - sessionCookie: []
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      description: >
        API key (prefix `lp_`) in `Authorization: Bearer <key>`. Works on all
        authenticated endpoints.
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
      description: >
        Same API key in the `x-api-key` header. Only consulted when no bearer
        `Authorization` header is present.
    sessionCookie:
      type: apiKey
      in: cookie
      name: llmproxy_session
      description: >
        Stateless signed browser session issued by the SSO callback or the
        admin password login. Carries the principal's role: valid on `/v1`,
        `/my`, and (admin role) `/admin/v1`. Non-GET requests with a
        mismatching `Origin` header are rejected with 403
        `cross_origin_rejected`.
...
paths:
...
  /auth/password:
    post:
      tags: [auth]
      summary: Admin password login
      description: >
        Signs the local admin principal in with the password from
        `LLMPROXY_ADMIN_PASSWORD` or the generated password file, and sets the
        session cookie. Available in local and SSO mode unless
        `LLMPROXY_ADMIN_PASSWORD_DISABLED` is set. Compared in constant time;
        a failed attempt is delayed 400 ms. Requests with a mismatching
        `Origin` header are rejected.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
              required: [password]
      responses:
        '200':
          description: Session cookie set.
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  role:
                    type: string
        '401':
          description: '`wrong_password`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: '`cross_origin_rejected`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: '`password_login_disabled`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/login:
    get:
      tags: [auth]
      summary: Start the SSO login flow
      description: >
        Sets a short-lived signed state cookie and redirects to the identity
        provider's authorization endpoint.
      security: []
      parameters:
        - name: retry
          in: query
          schema:
            type: string
            enum: ['1']
          description: >
            Set by the callback's automatic retry after a state mismatch;
            marks the new state so the retry happens at most once.
      responses:
        '302':
          description: Redirect to the IdP authorization endpoint.
        '404':
          description: '`sso_not_configured`: the proxy runs in local (no-SSO) mode.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/callback:
    get:
      tags: [auth]
      summary: OIDC redirect target
      description: >
        Validates the state cookie, exchanges the authorization code, fetches
        userinfo, applies the required-group gate and group-to-role mapping,
        upserts the principal (keyed on the IdP `sub`) and issues the session
        cookie.
      security: []
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
          description: IdP-reported error code, if the user denied consent.
      responses:
        '303':
          description: >
            Login succeeded (session cookie set, redirect to `/`), or the
            state did not match the cookie and the login is retried once via
            `/auth/login?retry=1` (a parallel login in another tab overwrites
            the single state cookie; the retried state is marked so a second
            mismatch reports `sso_state_mismatch` instead of looping).
        '400':
          description: >
            `sso_denied` (IdP reported an error), `sso_state_mismatch` (state
            cookie missing or mismatched after a retry) or `sso_code_missing`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: '`sso_group_required`: account is not in `LLMPROXY_OIDC_REQUIRED_GROUP`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: '`sso_not_configured`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: '`sso_exchange_failed` or `sso_userinfo_failed`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/models:
    get:
      tags: [openai]
      summary: List model aliases
      description: >
        OpenAI-shaped model list containing the aliases whose provider is
        enabled, sorted by alias.
      parameters:
        - name: endpoint
          in: query
          description: >
            Filter to aliases whose capability set contains this capability.
          schema:
            $ref: '#/components/schemas/Capability'
      responses:
        '200':
          description: Model list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelList'
        '400':
          description: '`invalid_endpoint`: unknown capability value.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/chat/completions:
    post:
      tags: [openai]
      summary: Chat completions (unary or SSE streaming)
      description: >
        Requires the `chat` capability; streamed requests additionally require
        `chat_stream`. Successful responses and upstream errors pass through
        with the upstream's status code and body. Proxy error codes:
        `invalid_json`, `model_required`, `model_not_found` (404),
        `endpoint_not_supported` (400), `request_too_large` (413),
        `provider_unreachable` (502), `internal_error` (500).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyRequestBody'
      responses:
        '200':
          description: >
            Upstream response, passed through. `text/event-stream` when
            `stream: true`, otherwise the upstream's JSON body.
          headers:
            x-llmproxy-provider:
              $ref: '#/components/headers/XLLMProxyProvider'
            x-llmproxy-model:
              $ref: '#/components/headers/XLLMProxyModel'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            text/event-stream:
              schema:
                type: string
        '400':
          description: >
            `invalid_json`, `model_required` (param `model`) or
            `endpoint_not_supported` (param `model`; message names the missing
            and supported capabilities).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: >
            `model_not_found` (param `model`): the alias does not exist, is
            or its provider is disabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '502':
          description: >
            `provider_unreachable`: connecting to or reading from the upstream
            failed (message names the failure class: timeout,
            connection_error, response_too_large).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          description: >
            Upstream error (status >= 400) relayed unchanged, with
            `x-llmproxy-error-source: upstream`.
          headers:
            x-llmproxy-provider:
              $ref: '#/components/headers/XLLMProxyProvider'
            x-llmproxy-model:
              $ref: '#/components/headers/XLLMProxyModel'
            x-llmproxy-error-source:
              $ref: '#/components/headers/XLLMProxyErrorSource'
  /v1/completions:
    post:
      tags: [openai]
      summary: Legacy text completions (unary or SSE streaming)
      description: >
        Requires the `completions` capability. Same passthrough behavior and
        error codes as `/v1/chat/completions`; note that `stream: true` does
        not require an extra capability here (only chat has a separate
        `chat_stream` capability).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyRequestBody'
      responses:
        '200':
          description: Upstream response, passed through.
          headers:
            x-llmproxy-provider:
              $ref: '#/components/headers/XLLMProxyProvider'
            x-llmproxy-model:
              $ref: '#/components/headers/XLLMProxyModel'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            text/event-stream:
              schema:
                type: string
        '400':
          description: '`invalid_json`, `model_required` or `endpoint_not_supported`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: '`model_not_found`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '502':
          description: '`provider_unreachable`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          description: Upstream error relayed unchanged.
          headers:
            x-llmproxy-error-source:
              $ref: '#/components/headers/XLLMProxyErrorSource'
  /v1/embeddings:
    post:
      tags: [openai]
      summary: Embeddings
      description: >
        Requires the `embeddings` capability. Always unary (the `stream` field
        is ignored). When `input` is a JSON array its length is capped at
        `LLMPROXY_MAX_EMBEDDING_BATCH` (default 2048).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyRequestBody'
      responses:
        '200':
          description: Upstream response, passed through.
          headers:
            x-llmproxy-provider:
              $ref: '#/components/headers/XLLMProxyProvider'
            x-llmproxy-model:
              $ref: '#/components/headers/XLLMProxyModel'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: >
            `invalid_json`, `model_required`, `endpoint_not_supported` or
            `embedding_batch_too_large` (param `input`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: '`model_not_found`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '502':
          description: '`provider_unreachable`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          description: Upstream error relayed unchanged.
          headers:
            x-llmproxy-error-source:
              $ref: '#/components/headers/XLLMProxyErrorSource'
  /my/keys:
    post:
      tags: [self-service]
      summary: Mint a key for the calling principal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  maxLength: 120
      responses:
        '201':
          description: Key created; the plaintext appears only in this response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyCreated'
        '400':
          description: '`invalid_json` or `invalid_label` (over 120 characters).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: >
            `cross_origin_rejected`: session-authenticated mutation with a
            mismatching `Origin` header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [self-service]
      summary: List own keys (metadata only)
      description: Returns up to 500 keys; not paginated.
      responses:
        '200':
          description: Key list.
          content:
            application/json:
              schema:
                type: object
                required: [keys]
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/KeyView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /my/relay-tokens:
    post:
      tags: [self-service]
      summary: Mint a relay token for the calling principal
      description: >
        A relay token attributes transparent-relay traffic to your principal.
        It is not an API key: it cannot authenticate against any other
        endpoint. Put it in the relay URL path, e.g.
        `ANTHROPIC_BASE_URL=https://proxy/transparent/anthropic/<token>`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  maxLength: 120
      responses:
        '201':
          description: >
            Relay token created; the plaintext appears only in this response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelayTokenCreated'
        '400':
          description: '`invalid_json` or `invalid_label` (over 120 characters).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: >
            `cross_origin_rejected`: session-authenticated mutation with a
            mismatching `Origin` header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [self-service]
      summary: List own relay tokens (metadata only)
      description: Returns up to 500 tokens; not paginated.
      responses:
        '200':
          description: Relay token list.
          content:
            application/json:
              schema:
                type: object
                required: [relay_tokens]
                properties:
                  relay_tokens:
                    type: array
                    items:
                      $ref: '#/components/schemas/RelayTokenView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /my/relay-tokens/{id}:
    delete:
      tags: [self-service]
      summary: Delete one of your own relay tokens
      description: >
        Deletion is the revocation mechanism: the row is removed and the token
        stops relaying immediately (404 `unknown_relay_token`). Usage history
        keeps the token id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: '`cross_origin_rejected` (session auth, bad `Origin`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: '`relay_token_not_found`: no such token on your principal.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /my/keys/{id}:
    delete:
      tags: [self-service]
      summary: Delete one of your own keys
      description: >
        Deletion is the revocation mechanism: the row is removed and the key
        stops authenticating immediately. Usage history keeps the key id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyDeleted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: '`cross_origin_rejected` (session auth, bad `Origin`).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: '`key_not_found`: no such key on your principal.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
  /my/usage:
    get:
      tags: [self-service]
      summary: Own usage summary
      description: >
        Aggregated by (model alias, endpoint) with request counts, cancelled
        counts, per-unit quantity sums and cost (null when unpriced).
      parameters:
        - $ref: '#/components/parameters/since'
        - $ref: '#/components/parameters/until'
      responses:
        '200':
          description: Usage rows sorted by alias then endpoint.
          content:
            application/json:
              schema:
                type: object
                required: [usage]
                properties:
                  usage:
                    type: array
                    items:
                      $ref: '#/components/schemas/UsageSummaryRow'
        '400':
          description: '`invalid_date`: `since` or `until` is not ISO 8601.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...